Code fix

+++ src/minecraft/net/minecraft/client/gui/GuiCustomizeWorldScreen.java
@@ -921,7 +921,7 @@ public class GuiCustomizeWorldScreen extends GuiScreen implements GuiSlider.Form
 
             if (f1 != null)
             {
-                f1 = f1.floatValue() + f;
+                f1 = f1 + f;
                 int i = guitextfield.getId();
                 String s = this.func_175330_b(guitextfield.getId(), f1);
                 guitextfield.setText(s);
